+++ CHANGES.md
@@ -10,6 +10,7 @@ Change Log
   * Removed the `direction` and `up` options to `Camera.flyTo`, which were deprecated in Cesium 1.6. Use the `orientation` option.
   * Removed `Camera.flyToRectangle`, which was deprecated in Cesium 1.6. Use `Camera.flyTo`.
 * Fixed an issue with transparency. [#2572](https://github.com/AnalyticalGraphicsInc/cesium/issues/2572)
+* Fixed improper handling of null values when loading `GeoJSON` data.
 
 ### 1.7.1 - 2015-03-06